re organize project file structure
